rareGoods/requirements.txt
asgiref==3.8.1
branca==0.7.2
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.4
django-autoslug==1.9.9
folium==0.17.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
pillow==10.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
xyzservices==2024.6.0
